requirements.txt
amqp==5.1.1
appdirs==1.4.4
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.6.1
Automat==20.2.0
billiard==3.6.4.0
CacheControl==0.12.6
cachy==0.3.0
celery==5.2.7
certifi==2020.12.5
cffi==1.15.1
chardet==4.0.0
cleo==0.8.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clikit==0.6.2
constantly==15.1.0
crashtest==0.3.1
cryptography==37.0.4
daphne==3.0.2
distlib==0.3.1
Django==3.2.14
django-active-link==0.1.8
django-celery-beat==2.4.0
django-modeltranslation==0.16.2
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
filelock==3.0.12
gunicorn==20.1.0
html5lib==1.1
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
jeepney==0.6.0
keyring==21.8.0
kombu==5.2.4
lockfile==0.12.2
msgpack==1.0.2
packaging==20.9
pastel==0.2.1
pexpect==4.8.0
Pillow==8.4.0
pkginfo==1.7.0
poetry==1.1.4
poetry-core==1.0.2
prompt-toolkit==3.0.37
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pylev==1.3.0
pyOpenSSL==22.0.0
pyparsing==2.4.7
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.1
redis==4.5.1
requests==2.25.1
requests-toolbelt==0.9.1
SecretStorage==3.3.1
service-identity==21.1.0
shellingham==1.4.0
six==1.16.0
sqlparse==0.4.2
tomlkit==0.7.0
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.7
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
XlsxWriter==3.0.3
zope.interface==5.4.0
django-environ
whitenoise
django-parler
social-auth-app-django
django-tinymce
django-parler-rest